Justering på Reservasjon og Utleiekontor

+++ src/no/hvl/dat109/Objekter/Innlevering.java
@@ -4,12 +4,12 @@ import java.time.LocalDate;
 
 public class Innlevering {
 
-    private int kredittkort;
+    private long kredittkort;
     private LocalDate returdato;
     private String regnr;
     private int kmStand;
 
-    public Innlevering(int kredittkort, LocalDate returdato, String regnr, int kmStand) {
+    public Innlevering(long kredittkort, LocalDate returdato, String regnr, int kmStand) {
         super();
         this.kredittkort = kredittkort;
         this.returdato = returdato;
